Don't attempt to get the object under the mouse after a click
Years ago, to work around missing focus events after a mouse click, we
attempted to get the object under the mouse and update the locus of
focus.

Obtaining the object under the mouse pointer has since become tricky
because 1) We only get the event in X; not Wayland. 2) The event we
get in X is in terms of absolute coordinates, but we must now use window
coordinates.

While we could take the approach that Mouse Review uses, namely use
ibwnck, we want to stop using libwnck; not use it even more. Therefore
under all of these circumstances, don't try try to get the object under
the mouse. If focus changes and we don't get an event, that's the app
or toolkit's fault and they'll need to fix their bug.