Speculative fix for flat review sometimes failing in terminal
It appears that there is not always a line at offset for the last or
second-to-last characters in terminal. Detect this condition in AXText's
find_last_visible_line. Make a similar change just in case for its
find_first_visible_line.

Also add more verbose debugging for a related issue in which flat review
fails in the terminal due to an ancestor of the terminal claiming to have
no children event after we recursively clear the AT-SPI2 cache. We cannot
due anything more on our end, but flagging the issue in the debug output
so we spend less time hunting it down in the future seems useful.