Event Manager: Add means to pause and clear the event queue
When shutting down, we want to stop processing events and not accept
any additional ones. In order to facilitate this, we deactivated the
Event Manager first, then the Script Manager. But the Event Manager
is the one which manages our Atspi.Device instance and sets it to None
upon deactivation. Thus when the Script Manager attempts to remove key
grabs, the Atspi.Device is already gone.

While this does not appear to have any real-world consequences (a la
permagrabs), we should properly clean up after ourselves during shutdown.
Therefore:

* Create a means to pause and clear the event queue independent of
  Event Manager deactivation
* Use it prior to deactivating the script manager
* Finally deactivate the event manager after doing so for the script
  manager