Add CLI option for specifying a speech system
Orca will soon have support for Spiel, in addition to SpeechDispatcher,
so add an option to specify the speech system from the command-line.

Collect a list of successfully loaded speech system modules, then check
the provided name against those. If the request system is unavailable,
the list of available systems will be included in the error message.

closes #460