Web: Preserve context when focus removes one of its children
handleEventForRemovedChild does not do any recovery if the removed
child is not the current focus nor an ancestor of the current focus.
In that case, the existing context should still be valid. If we
clear the context but do no recovery, we'll lose our place and jump
to the top. One example where this can be seen on sites where a
widget has an associated tooltip. When we move to it, the tooltip
is popped up. When we move away from it, the tooltip gets removed.
That child removal should not cause us to clear the context; just
the cache.