Web: Fix skipping over focusable, named text-block elements with empty children
Example: Some StackExchange instances have a focusable "note"/comment role
with a name (e.g. "Accepted"), and a single child div which is empty.

Because we treat text block elements as something caret navigable if they
have text, we were entering the element, finding the empty child div and
moving on to the next offset beyond the note/comment. Fix this by checking
focusable, named text-block elements for useful/navigable children. If we
don't find any, treat the element as if it had no children.