Web: Don't use default script's updateBraille functions if "\ufffc" in text
We were always using the default script's updateBraille functions in
focus mode. As a result, a link in a message being composed in Thunderbird
was showing up as an embedded object character. Fix that by checking if
an embedded object character is present in the text to be displayed. If
so, let the web script handle it.