Remove the unused VCL.py
This was a script with no functionality whatsoever. It simply subclassed
the default script. It was created for StarOffice back when it (and maybe
LibreOffice) reported VCL as its toolkit. But LibreOffice claims Gtk as
its toolkit, and the soffice script has subclassed the default script
directly for years. Therefore delete VCL.py