Remove theoretically-unused import of pyatspi
We are officially pyatspi-free. Please note, there may be a case or two
where we actually are not. This is due to the fact that the majority
of pyatspi API does not have the string "pyatspi" in it. With the
import in place, both AT-SPI2 calls and pyatspi calls worked. Without
the import in place, any pyatspi calls we missed will result in an
error.

See issue #300.