Add a speechserver for Spiel
Spiel is a new a speech synthesis API for desktop Linux and others,
allowing clients to easily perform TTS with a variety of synthesizers
planned including eSpeak, Piper, and mimic3.
Add SSML helper class

Move SSML utilities into a separate import so they can be re-used,
with prelimary support for additional SSML features.

The README has been updated with safe instructions for getting
started with the initial implementation of Spiel, highlighting that
speech-dispatcher is the default TTS service.

closes #452