Ignore selection-changed events from trees outside of the active window
If a mail client fires a selection-change event for a mailbox or message
list, and the user is in a different window reading a message, we don't
want to update focus outside of the message being read. Doing so can
cause us to stop presenting the message until something causes the
message to reclaim focus.