Braille: Remove _generateEol overrides; place all logic in default's
We had very inconsistent use of the braille end-of-line indicator.
Current feedback on the list is that it makes sense to include it at
the end of all editable lines as well as at the end of a line we know
is code/preformatted text. There was also an indication that making
things configurable would be nice.

This commit is the first step of the first item: It removes the
web and webkitgtk scripts' override of _generateEol and places the
agreed-upon rules (end of all editable or code/pre-formatted lines)
into the default's _generateEol.

The second step of the first item is to go through all of formatting.py
and remove the role-specific inclusion (or exclusion) of the indicator.
That will be done in a subsequent commit.

Making it fully configurable is on the TODO list.