AXText: Handle offset of -1 in the *_at_offset functions
Some implementations return an offset of -1 if the caret is not in the
text object. Currently we are treating that as an error condition,
leaving it up to all the individual callers to correct (e.g. via
`max(0, offset)`). It would be easier and simpler to do the correction
in AXText.