Speculative fix for SMPlayer's bogus screen-based bounding box
Treat negative width or negative height as invalid so we don't try to
see if it fits in the window's bounding box. That bug should be fixed
in the player or Qt. But we don't want to regress the user experience
by expecting valid coordinates from objects in flat review.

See issue #469.