Action Presenter: Clean up after menu is hidden
If the action presenter menu is dismissed and the underlying app does not
fire a window activate and/or focus event, we won't update our location
and active script. This can cause things like caret navigation and
structural navigation to not work. Fix this by doing the following:

* Connect to the "hide" signal of the action-presenter menu
* When the menu is hidden, clean up by clearing the focus manager's
  state, setting the active script, the active window, and focus.

This should cause us to present the restored location as well as
refresh keygrabs.