Remove the notify-osd script
(Note that this is NOT the same thing as notification-daemon which is used
in MATE.)

Looking at the notify-osd page https://launchpad.net/notify-osd, the last
release was nearly 12 years ago. And the last Ubuntu package it was shipped
in is Xenial. Those environments are way too old to run the current version
of Orca. Therefore it should be safe to remove this script.