Remove hasMatchingHierarchy hack from Pidgin script
hasMatchingHierarchy is extremely fragile. We were using it as part
of a heuristic to identify if a new chat tab was opened in the
background of the current chat window. The part of the heuristic
to identify the window doesn't need the hierarchy. And we can tell
if the tab is in the background because it's not selected and its
name doesn't match the frame's name.