Replace chat scripts' use of hasMatchingHierarchy
hasMatchingHierarchy is extremely fragile. We were using it for a
heuristic identifying buddy lists. We can accomplish the same
thing by looking for ancestors with the sorts of roles we see in
the existing apps' buddy lists.

This was the last use of hasMatchingHierarchy, so delete that
function from the script utilities.