Fix word echo for Writer; merge gnome-shell word echo checks into default
Writer returns "", 0, 0 for the current character when the caret offset
is at the end of the text. Handling that scenario was removed as part of
code clean up. Restore that check. Also merge the workaround for gnome-
shell, which returns an offset of -1 when the caret offset is at the end
of text, into the default script's echoPreviousWord.