Don't treat "system" text deletion events as delete commands
When the user deletes text, we want to present the text at the new
location. Unfortunately, when the user deletes a message in Thunderbird,
Thunderbird fires text deletion (and insertion) events that appear to be
it re-using the current list/tree item. This causes us to speak one or
a few characters unexpectedly.

We do not want to handle these deletions as user-triggered text changes.
Fortunately, we can distinguish these deletions from typical deletions
because Gecko appends the "system" suffix in this case. Use that in
isDeleteCommandTextDeletionEvent to filter out these events.