Thunderbird: Ensure all name-changed events get handled
Changes in Thunderbird are impacting presentation of deleted items.
Before, the deleted item would actually be deleted. Then we'd get
a selection change event and present it. Now the events seem to be
as follows:

1. The user presses delete
2. Thunderbird changes the name of the deleted item to reflect the
   newly-selected item (and fires an event)
3. Thunderbird fires the selection-change events

Given the above, Orca gets a selection event as expected, but does
not present it because the selected accessible object has not changed.
Thus we need to present the name-change instead.

Thunderbird was returning early for name-change events that were not
from the document in relation to a deletion. So stop returning early.
In addition, the web script was always returning True (event was
handled). That's a bug. Return False if the event wasn't handled.