Don't present name changes of frames if the name comes from the editable focus
As a side effect of 621d4efa61257e55b8ec651abb3dc959c5accf58, Orca got
super chatty in Thunderbird whenever one was typing the subject of an
email. This was caused by our no longer ignoring most name-change events
from Thunderbird due to changes in Thunderbird (reusing objects and
changing their name rather than creating new objects).

In order to solve this particular regression in the user experience,
ignore name changes that are fired on a frame if focus is in an editable
field and the complete text of that editable field is contained within
the new name of the frame.