treat error in get_text_attributes_at_offset as no attributes everywhere
When GetAttributeRun, for example, is not implemented, or
fails otherwise for a text, we know of no attributes for
it, and do so for the complete content of it.

Thus, it makes sense to claim said range – the complete
text – has no attributes.

That makes users like splitSubstringByLanguage still use
the string, even though the attributes are not known.

Signed-off-by: Markus Göllnitz <camelcasenick@bewares.it>