Web: Treat name changes to frames as browser UI noise
Prior to 621d4efa61257e55b8ec651abb3dc959c5accf58, we were treating
(incorrectly) all name change events as handled, and functionally
ignoring those events. Now that we no longer do so, Orca is getting
chatty. One instance of such chattiness is when fast-forwarding on
YouTube. YouTube briefly changes the name of the frame each time
shift+right is pressed and changes it back when released. We should
not be presenting those changes.

Rather than trying to handle specific instances of name changes on
browser frames, it should be safe to ignore them: When a new page is
loaded, we present the name of that page which typically matches the
frame name. Plus back when we were ignoring all name changes, no one
reported Orca failing to present the changed name of the current window.