Event Manager: Always set the active script in response to window:activate
We were only doing so when the event was for the active window. However,
we sometimes correct the active window in response to an input event, so
it is possible to get a window:activate event for a window that we set
as our active window. The script associated with that window should be
used for processing object events.