Focus Manager: Update the active script in set_active_window
It is possible for something to update the active window without updating
the active script. 8325d8af corrected for the one known instance of this,
but it should not be possible for this to happen in the first place.

Have set_active_window update the active script based on the current
focus (if it's in the active window) or active window.