Bypass Mode: Disable when script is deactivated
If Bypass Mode is toggled on and the user switches applications,
we were winding up in a strange state where keygrabs for the new
app were activated and working but bypass mode thought it was still
active. To prevent this weirdness, automatically disable bypass mode
when a script becomes inactive.