Remove _speakTextSelectionState hack
_speakTextSelectionState spoke messages that were often confusing as
well as untrue. For instance, if Ctrl+Shift+PageDown were pressed and
we got a text selection changed event, if text was selected Orca would
say "line selected to end from previous cursor position". Parsing what
that means is hard, and we did no checks to see if that condition was
really what happened.

It does not make sense to speak unverified messages about the selection
state based on keyboard shortcuts that presumably were in place in Gtk2.
Therefore, stop doing this. We can consider providing more useful, and
true, messages (e.g. "10 lines selected") as a follow-up change.