Remove eventIsUserTriggered hack
This hack was meant to filter out window-deactivate events that were not
user triggered, such as a temporary popup from the window manager. The
logic was based on the time of the last input event and assumed that if
there was no last input event, then the window deactivation was not user
triggered and could be ignored. Now that we can no longer count on being
notified of input events, this hack is even more fragile. The real fix
is to either get these events stopped, or to be given a means to identify
that a temporary popup is responsible for them so we can reliably filter
them out.