Web: Defer to default script's presentObject() when the object is the document
The web script's presentObject() calls getObjectContentsAtOffset() to determine
what to present for that object. Normally that results in a list of one or two
content items to present. But when the object to be presented is the document
itself, the resulting content items will likely be in the hundreds (at least).
It's like doing a SayAll, but gathering the full content prior to saying anything.
This makes it appear that Orca has hung.