Have is_message_dialog return early when possible
Thanks to the collection interface, it is possible to very quickly
look for any widgets that suggest a dialog is not a simple message
dialog. Unfortunately, in some apps (e.g. Pidgin, a Gtk2 app) this
seems to make Pidgin crash for reasons still to be determined. In
order to try to avoid this sort of crash, use several collection-
based checks to quickly locate the first container which suggests
that a dialog is not a message box. Hopefully this will prevent
AT-SPI2 from digging so deeply in the tree that Pidgin crashes,
and it's a reasonable change to do anyway.

See issue #485.