Bump Spiel requirement to 1.0.1
Ensure the build fails if attempting to build Orca with support
for an older version of Spiel. The previous version contains a
bug that may throw a pre-condition critical error that can not
be caught in language bindings.

In addition, update the README to purge the existing subprojects
which is safe and guaranteed to use the correct wrap version.