Remove orca_state.capturingKeys
This state is no longer needed due to how we handle keybindings
and grabs. In particular, when the preferences dialog set it to
True, it also called removeKeyGrabs() which in turn not only
removes the grab, but it also clears the keybindings. As a result,
when the input event comes in, the script has no bindings for it
so we treat it as not a command.

See issue #488.