gnome-shell: Remove obsolete hack/workaround
We were checking and filtering out focus events from a window which
could not be the active window. This bug does not appear to be present
any longer. If it turns out the gnome-shell issue is still present,
the correct place to ignore that event is in the event manager's
_ignore function.