Ignore focus events from objects with the "window" role
Commit 29f45293f2e removed an old hack for bogus focus events being
fired from the gnome-shell script. It turns out that hack is still
needed. For instance, when Alt+Tabbing, the switcher window will
emit a focus event after the window switch occurred. This can cause
us to update the locus of focus, change keygrabs, etc.

Restore the hack, but not in the gnome-shell script. The event manager
is where all logic related to ignoring (or not) object events belongs.