Persist runtime settings between terminal and terminal app scripts
Scenario: speech is disabled in the mate-terminal app preferences.
The user is in the terminal and presses Orca+S to temporarily enable
speech and then enters the File menu. We want speech to remain enabled.
It was not. The reason why is that speech was enabled while the terminal
script was active. Going into the File menu causes the mate-terminal
script to become active and its settings loaded, stomping on the run-
time-enabled speech. To solve this problem:

1. Add means to store and retrieve run-time settings to the settings
   manager. These are cleared whenever app settings are loaded.
2. In the script manager's set_active_script, retrieve the run-time
   settings when the app associated with the old script is the same
   as that of the new script. Having done so, then activate the new
   script and set the stored run-time settings.
3. Move the calls to checkBrailleSetting() and check_speech_setting()
   out of script.activate and into set_active_script. This is needed
   because at the time the new script is activated, we don't have
   full knowledge of whether or not speech and/or braille should be
   shutdown or started up.