Braille: Adjust caret offset of -1 to start offset when provided
As a side effect of removing getTextLineAtCaret, sometimes braille
was not being displayed for Chromium. This appears due to a number
of factors:

* We were attempting to display text for the document itself
* Chromium's native navigation was not enabled by the user
* Even though we attempt to set the caret offset, Chromium does not
  always respect that. And apparently when its native navigation is
  disabled, it returns a caret offset of -1.
* AXText's get_line_at_offset can be called without a caret offset,
  in which case it uses the reported offset -- or 0 if the offset
  is negative.
* The line of text for the document object at offset 0 is almost
  always an embedded object character, which we convert to a space
  when displaying braille.

In order to fix this, adjust the caret offset from -1 to the start
offset (which the web script always provides) and use that offset
rather than the default when calling get_line_at_offset.