Convert WebKitGTK support to be based on the web script
Please note: There are a number of known issues -- and possibly some
undiscovered issues -- that will need to be fixed in WebKitGTK before
this new support works as expected. See, for instance:

* https://bugs.webkit.org/show_bug.cgi?id=268154
* https://bugs.webkit.org/show_bug.cgi?id=270354

The implications of the two combined issues include:

1. Orca might not present the content of a newly-loaded document.
2. Orca might not present certain embedded objects when presenting
   a line of content (e.g. it might not say the link text).
3. Caret navigation might fail completely.

Depending on the content and application, using one or both of the
following techniques can sometimes be used to work around the above
issues:

1. Tab to move focus to a link or other object. Doing so will put Orca
   in the document.
2. Switch to focus mode and enable native caret navigation with F7 in
   apps that support this (e.g. Yelp). This is needed any time caret
   navigation fails to move to the desired location (up, down, left,
   right, home, end, etc.).

Note that structural navigation commands seem to work as expected;
it's reading content in browse mode that is most problematic.