Web: Don't require the active state from ARIA dialogs
We get window:activate events from ARIA dialogs but those "windows"
will almost certainly lack the active state. That's never been a
problem because best practice is for the author to set focus to a
widget inside the dialog, and the resulting event causes us to present
the newly-focused widget along with the new context (i.e. the dialog
itself). When best practice is not followed, we'd ignore the event
and remain silent due to nothing claiming focus.

Because we cannot count on best practice being followed, set our
focus to an ARIA dialog in response to a window:activate event.
This should cause us to present the dialog, and only do so once,
regardless of whether or not the expected focus event comes.

See issue #495.