Eliminate _generateTableCell2ChildLabel, _generateTableCell2ChildToggle
These functions landed in 2007 to work around a specific issue in
GailTreeView and/or an app that used it. Removing these functions
does not appear to have any impact in the gtk-demo tree view demos
for Gtk2 or Gtk3, so we are potentially doing extra work for no good
reason. If it turns out this workaround is still needed, we can do a
cleaner version of it.