Remove old LibreOffice logic related to combo boxes
This logic predates the pre-Gtk3 LibreOffice UI and was indended to work
around oddities/bugs in LO's custom comboboxes. The non-Gtk comboboxes
that remain (e.g. on the formatting toolbar) are no longer comboboxes;
they are panels with descending tables. As a result, the logic related
to comboboxes which expected a particular role and accessibility tree
no longer applies. Therefore delete the code in question. Dealing with
the new LO combobox structure can be done as a follow-up change.