Remove on_focus from sleepmode script; add debugging to default
The on_focus handler for sleepmode was just printing a debugging line
because no object events are processed in sleepmode. The on_focus
handler in the default script already does nothing. But now that we're
removing on_focus handlers from other scripts, it would be nice to see
a logline regarding why we're ignoring this event (which was deprecated
a decade ago).