Remove on_focus from terminal script
While the missing event (which caused the workaround to be added) is
still be a bug, removing the handler does not appear to regress the
behavior that workaround addressed. This may be due to the fact that we
are now clearing the AT-SPI cache when setting the locus of focus and/or
changes in how we process text changes (i.e. other workarounds are
kicking in making the on_focus one redundant).