Web: Make fix in 5977142ec more conservative
Sometimes in caret navigation the prior object will be the same as the
object we're generating speech for. 5977142ec was meant to address
name repetition in editable content, so adjust that check to only
apply to editable content. This fixes a regression in which the name
of buttons was not being presented during caret navigation.