Speculative workaround for missing focused-changed events
The Pluma text view is not firing object:state-changed:focused events when
a user opens a file via Caja from within MATE. This results in two things
happening: 1) We don't present the newly-focused text view because we don't
know it got focus, 2) we don't present subsequent caret-moved events when
the user arrows in the text view because it claims to be unfocused (i.e.
and not something the user is in).

As a speculative fix for user testing, try clearing the AT-SPI2 cache for
all caret-moved events originating from Gtk widgets which claim to not be
focused prior to letting the default script's logic do its thing.

See https://gitlab.gnome.org/GNOME/gtk/-/issues/454.