Web: Ignore caret-moved events resulting from tabbing to a link
In user agents where tabbing to a link causes a caret-moved event to be
emitted prior to the focused event, we won't present the latter because
we silently updated our location due to the former. Currently this only
seems to happen with WebKitGtk, but it theoretically could happen in any
user agent.