Eliminate the use of formatting strings for the sound generator
Note that the sound generator was a feature-in-progress that got
put on hold. Therefore there should be no functionality change
as a result of this commit. After some additional refactoring has
been completed, we can resume work on generating sounds for objects,
events, and mode changes.

This commit also removes the web script's sound generator. At the
time of creation, it seemed to make sense, but in looking at the
code now, its functionality belongs in the default script's.

See issue #496.