README: Add Qt sample for sending announcement event
Similar to the existing GTK 3 and GTK 4 sample apps,
add a small sample Qt app that demonstrates how to
send a QAccessibleAnnouncementEvent, which is bridged
to an AT-SPI "Announcement" event on Linux.

Qt API and bridging to AT-SPI were added in Qt
commits [1] and [2] and will be contained in Qt >= 6.8:

    commit 010952a55ee9ca9731f8ad15f2773d0bae645840
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Tue Apr 9 16:43:21 2024 +0200

        a11y: Add new QAccessibleAnnouncementEvent

    commit 24986ccb7b2020cbb574700a31667e88ed243194
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Tue Apr 9 16:17:58 2024 +0200

        a11y atspi: Forward new QAccessibleAnnouncementEvent to AT-SPI

The corresponding Qt issue/feature request [3] contains a C++ sample
app that can be run as is with the current qtbase git dev branch.

In order for the Python sample added to the README to work,
PySide (the Qt Python bindings) also needs to be updated.
Pending Gerrit change: [4]

[1] https://code.qt.io/cgit/qt/qtbase.git/commit/?id=010952a55ee9ca9731f8ad15f2773d0bae645840
[2] https://code.qt.io/cgit/qt/qtbase.git/commit/?id=24986ccb7b2020cbb574700a31667e88ed243194
[3] https://bugreports.qt.io/browse/QTBUG-75003
[4] https://codereview.qt-project.org/c/pyside/pyside-setup/+/561134

Fixes: #447