_generateLabelAndName() should prefer largest redundant string
_generateLabelAndName() has logic to check if the label and name are
redundant to one another. When that is the case, it was returning
the label and ignoring the name. What should happen is that when only
the label or the name should be presented, the longer one is chosen.
This commit makes that change.

It also fixes a regression in Discord in which messages are articles
with a detailed name and a less detailed label. The regression resulted
from the elimination of _generateLabelOrName which, in the case of the
web script, preferred the accessible name.