Evolution: Ignore events coming from the document preview
When focus is is one of the trees and the user is arrowing up/down,
we see the following events that are incorrectly causing us to
update our location:

* busy-changed from the newly-loaded document preview
* caret-moved from the newly-loaded document preview
* focused-changed from some iframe I cannot find in the accessibility
  tree, and wasn't motivated to debug at this time

Ignore the last one because iframes, event those one can find, should
not fire focus events; the document should. Ignore the first two if
we're in a non-WebKitGtk tree or tree table and the user just arrowed.