_generateLabel should return nothing for accessible web content
User agents expose labels (from HTML, ARIA) as the accessible name as
per the specs. Thus there is normally no need to use _generateLabel
for web content. The one exception to this is unlabelled form fields
for which we have heuristics to infer the functional label.

If, in the case of accessible content, we allow the default script's
_generateLabel to kick in, ultimately the redundant nature of the label
to the name will be caught and filtered out. But why do the extra work?

In addition, there are scenarios where some of the labelling objects
lack names and have embedded object children in there accessible text.
When this occurs the logic that checks for redundancy can fail causing
us to concatenate the label(s) we kept to the name and double-speaking
bits and/or speaking things out of order.

Just return nothing for accessible web content to solve both issues.