Don't insist on showing+visible for combobox selection changes
In certain combobox implementations (e.g. Thunderbird's), when the
combobox is collapsed its menu child will lack state showing and
state visible but fire selection-changed events. While we normally
want to ignore selection-changed events from objects that are not
showing, one could make a reasonable case that the collapsed menu
is not showing or visible and is the object which should be firing
the event. Therefore, handle it on our end.