Eliminate chattiness from redundant objects in ancestry
Example: Adwaita's List Demo has a panel named "Lists" which is inside
another panel named "Lists" which is inside a third panel named "Lists".
We don't want to say "Lists panel" three times; only once.

For now, treat an ancestor as redundant if it has the same name and role.
Additional refinement might be needed in the future based on what we find
in the wild.