Don't descend lists or list items when looking for unrelated labels
Looking at the Adwaita Preferences demo, there are list items with
descendant labels. These labels serve as the name of their ancestor
item. There should be something relating the label to the item. That
said, unrelated labels are typically in things like panels and pages;
not in lists and list items. So add them to the group of roles we
don't look in.