Only speak role of ancestor panel if the current object is a widget
Generic Adwaita status pages seem to consist of a named, described
panel with a descendant scroll pane which claims focus. Saying
"panel scroll pane" is rather chatty given both are containers.
Saying "panel" is still useful to identify a group of widgets,
therefore only speak that role if the current object is a widget.
(We still say "scroll pane" because the user should know what the
focused item is.)