Reject children-changed:remove events from gnome-shell sooner
gnome-shell periodically floods us with children-changed:removed events,
e.g. when exiting the app switcher. We normally ignore these events --
unless the locus of focus is dead. Checking for the locus of focus being
dead allows us to attempt to recover when an app destroys the object we're
on. But such destruction tends to occur in web pages/apps; not gnome-shell.
And queuing up all those events from gnome-shell only to ultimately ignore
them later is a non-performant waste of time.