Set priorObj to newly-active window when old focus is null
We determine what ancestors to present based on what changed from
the previous focus to the new focus. However, if the previous focus
got destroyed (which could happen when using the gnome-shell switcher)
and we don't get the expected window:activate event (which seems to
happen with either Gtk4 and/or Adwaita), we don't generate ancestors
because we don't know what changed and don't want to be too chatty.

Work around this condition by setting the old focus to the newly-
active window so that we generate the full ancestry -- like we would
have done had we gotten the expected window:activate event.

See issue #508.