Remove LO custom _generateLabel()
The LO _generateLabel() differed from the default's by falling back on
the parent's label. The reason why is not clear, but it seems to have
been connected to the Where Am I logic which was moved into the speech
generators back in 2009.

The side effect of the LO fallback is that when focus/selection changes
within a labelled parent, such as the list of recent files when no
document is open, we incorrectly repeat that parent every single time.
Therefore remove this old fallback. If we discover a scenario for which
it is needed, we can always address that need in the appropriate code.