Fix bug in on_caret_moved
If the caret moves in a focused object that is not the locus of focus,
we attempt to correct for that condition by updating the locus of focus
to the event source. However, subsequent code then compares the event
source to what was the previous locus of focus. Fix that by updating
the value of focus prior to doing the comparison.