Fix regression in role presentation
Commits d9a3ae8b0e and d70919051ef0 eliminated unwanted presentation
and repetition of text roles (article, form, paragraph, and section).
Unfortunately that regressed the presentation of certain roles during
strutural navigation. The reason is that _generateRoleName includes
logic to present certain ancestor roles (heading, link, etc.) as if
they belonged to the text object.

Ultimately this should be handled with a dedicated generator, and/or
a treat-as-role override. For now handle it by adding the offending
roles to the lists of roles we don't generate even though we asked
for role generation.