Mouse review: Treat no display as not mouse-review capable
At least one user has reported that enabling mouse review under Wayland
causes Orca to not speak anything. I cannot reproduce that issue. There
is an error from Wnck about not working in X11, but things continue to
work (other than mouse review).

This is a speculative, hopeful fix for the user's reported issue, namely
if we can import Wnck successfully but cannot get the default screen,
treat it as a failure so that no mouse-review logic is attempted.