Remove "Speak multicase strings as words" setting
This is an old setting that, if enabled, caused us to convert text
such as "HelloWorld" into "hello world" before sending it to the
speech engine for the purpose of having the text spoken correctly.
Testing suggests that the speech synthesizers are already handling
multi-case strings as we'd like, so there's no need for us to do this
additional work.