Remove the outdated gcalctool script
This minimal script is doing nothing for the current version of
gnome-calculator and is not being used for anything else. The limited
functionality it had was based on a non-longer applicable heuristic to
find the display in the accessibility tree. And that was only needed
if the user was interacting with the buttons for input.

Now that we have the accessible announcement event, calculator apps
can use that to tell Orca what to present in response to a calculation
when the editable display is not being used. Bugs to the effect will
be filed.