Web: Fix regression in which entry fields are skipped in browse mode
Commit 4b71c07bccbf31 included a change to always treat editable
objects as text objects regardless of their contents. As a side
effect of this change, find{Next/Previous}CaretInOrder skipped
over empty entries. Fix that by only descending text objects that
actually have text. In the case of an empty entry, this will cause
subsequent logic to treat the entry as the next/previous caret
position.