Live Regions: Attempt to make things better
ARIA live region support has always been challenging because user agents
fire both text-changed and children-changed events, and they do so both
in the case of there being a live region to present and in the case of
a DOM mutation. Furthermore, in the case of (at least) Chromium, the
children-changed events don't always come in presentational order and
duplicate events are sometimes fired.

Based on some functional testing, we seem to be getting text-insertion
events reliably and in the correct order. Therefore in order to try to
make things better:

1. Only handle text-insertion events as presentable live region events.
2. Store the previously-presented message along with a timestamp and
   use this to weed out duplicates. Do this in the live region manager.
3. Remove code from the web script which attempted to filter out
   duplicate code

This change seems to make presentation of Google docs content with
links complete (Firefox) and less chatty (Chromium). It also seems
to work as expected with some sample live-region test pages.

Let the bug reporting begin!