Add an offset argument to synthesized char events
When routing/clicking to the current character, by default we've used
the actual text offset. That doesn't work in flat review because the
current item being reviewed almost certainly is not the same location.
The flat-review code was therefore handling this by synthesizing the
event on coordinates it determined. Enabling callers to specify the
character offset of interest simplifies the code and also lets us
eliminate route_to_point, which is needed to begin using the new(ish)
atspi_device_generate_mouse_event.