Handle Chromium's differing behavior for line at offset at end
Some return the last line when we ask for the line at offset when
offset is the character count. Others (at least Chromium) does not.
As a result, in VS Code, if one does not press return at the end of
the final line, sets the caret to the end of the line, and presses
Up followed by Down, Orca was not presenting the final line as
expected. Handle that case.