Event Manager: Ignore events from apps unknown to AT-SPI2
Gecko apps are firing events from two different accessible apps:
1) The actual accessible app, known to AT-SPI2's registry, whose
   objects report the "Gecko" toolkit.
2) Some other accessible app, unknown to AT-SPI2's registry, whose
   objects report the "Gtk" toolkit.

We have logic in place to try to handle this when processing the
accessible events, but it doesn't appear to be enough. As a result,
we're updating keygrabs unnecessarily, repeating window titles, and
occasionally updating focus to some panel because it insists it is
focused.

See https://bugzilla.mozilla.org/show_bug.cgi?id=1882794

Since that issue is not getting any attention, ignore all events
from apps unknown to AT-SPI2. This seems to make things better and
without any known (so far) side effects.